Refactoring buttons + rename files
